+++ libs/viz/src/lib/image-download/image-download-config.ts
@@ -3,7 +3,6 @@ import { VicImage } from './image-download-enums';
 import { ImageDownloadOptions } from './image-download-options';
 
 const DEFAULT = {
-  fontEmbedCSS: undefined,
   backgroundColor: '#fff',
   quality: 1,
 };
@@ -12,8 +11,7 @@ export class VicJpegImageConfig implements ImageDownloadOptions {
   backgroundColor: string;
   containerNode: HTMLElement;
   fileName: string;
-  filter: (domNode: HTMLElement) => boolean;
-  fontEmbedCSS: string;
+  filter: (domNode: Element) => boolean;
   imageType: string;
   quality: number;
 
@@ -28,8 +26,7 @@ export class VicPngImageConfig implements ImageDownloadOptions {
   backgroundColor: string;
   containerNode: HTMLElement;
   fileName: string;
-  filter: (domNode: HTMLElement) => boolean;
-  fontEmbedCSS: string;
+  filter: (domNode: Element) => boolean;
   imageType: string;
   quality: number;
 
@@ -44,8 +41,7 @@ export class VicSvgImageConfig implements ImageDownloadOptions {
   backgroundColor: string;
   containerNode: HTMLElement;
   fileName: string;
-  filter: (domNode: HTMLElement) => boolean;
-  fontEmbedCSS: string;
+  filter: (domNode: Element) => boolean;
   imageType: string;
   quality: number;